requirements.txt
apiclient==1.0.4
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
http-client==0.1.22
httplib2==0.14.0
idna==2.8
oauth2client==4.1.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycurl==7.43.0.3
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
uritemplate==3.0.0
urllib3==1.25.6
youtube-dl==2019.11.5
